third_party/WebKit/LayoutTests/http/tests/security/calling-versus-current.html

<!DOCTYPE html>
<html>
<head>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <script>
        if (document.location.hostname == "127.0.0.1") {
            document.location.hostname = "subdomain.example.test";
        } else {
            var t = async_test("Calling context doesn't change via document.domain.");

            t.step(function () {
                var i = document.createElement('iframe');
                i.src = 'resources/innocent-victim.html';
                i.onload = t.step_func(function () {
                    window.f = frames[0].atob;
                    document.domain = 'example.test';
                    assert_equals(btoa(window.f('PASS')), 'PASS');
                    t.done();
                });
                document.body.appendChild(i);
            });
        }
    </script>
</body>
</html>
